requirements.txt
python-magic
#python-magic-bin -> windows
